finalize user management feature
