+++ issuetracker-client/package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
+    "@types/node": "^20.12.11",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
