+++ issuetracker-client/src/App.module.css
@@ -1,0 +1,3 @@
+.navbar {
+    box-shadow: var(--shadow-2);
+}
