+++ issuetracker-client/src/App.tsx
@@ -1,18 +1,21 @@
 import "@radix-ui/themes/styles.css";
+import styles from "./App.module.css";
+import "./theme.css";
 
-import { Container, Flex, Theme } from "@radix-ui/themes";
+import { Flex, Section, Text, Theme } from "@radix-ui/themes";
 import { Link, Outlet } from "react-router-dom";
 
 function App() {
     return (
         <Theme>
-            <Container size="1">
-                <Flex gap="4">
+            <Section p="3" className={styles.navbar}>
+                <Flex gap="6" justify="start">
+                    <Text>Issue Tracker</Text>
                     <Link to="/users">Users</Link>
                     <Link to="/teams">Teams</Link>
                 </Flex>
-                <Outlet />
-            </Container>
+            </Section>
+            <Outlet />
         </Theme>
     );
 }
