+++ issuetracker-client/src/modules/teams/styles/add-user-dialog.module.css
@@ -1,0 +1,20 @@
+.dialog {
+}
+
+.list {
+    margin-bottom: var(--space-4);
+    height: 280px;
+    overflow-y: auto;
+}
+.list-item {
+    padding: var(--space-2);
+    cursor: pointer;
+}
+
+.active {
+    background-color: var(--gray-3);
+}
+
+.selected {
+    background-color: var(--indigo-4);
+}
