+++ sql/sqlite/schema.sql
@@ -1,4 +1,8 @@
-CREATE TABLE IF NOT EXISTS account (
+DROP TABLE IF EXISTS team_member;
+DROP TABLE IF EXISTS team;
+DROP TABLE IF EXISTS account;
+
+CREATE TABLE account (
 	id text PRIMARY KEY,
 	username text UNIQUE NOT NULL,
 	first_name text,
@@ -6,16 +10,16 @@ CREATE TABLE IF NOT EXISTS account (
 	date_created text
 );
 
-CREATE TABLE IF NOT EXISTS team (
+CREATE TABLE team (
 	id text PRIMARY KEY,
 	name text,
 	date_created text NOT NULL
 );
 
-CREATE TABLE IF NOT EXISTS team_member (
+CREATE TABLE team_member (
 	id integer PRIMARY KEY AUTOINCREMENT,
-	team_id text REFERENCES team (id) NOT NULL
-	    ON DELETE CASCADE ON UPDATE CASCADE,
-	account_id text REFERENCES account (id) NOT NULL
-	    ON DELETE CASCADE ON UPDATE CASCADE
+	team_id text REFERENCES team (id) 
+	    ON DELETE CASCADE ON UPDATE CASCADE NOT NULL,
+	account_id text REFERENCES account (id)
+	    ON DELETE CASCADE ON UPDATE CASCADE NOT NULL
 );
